+++ themes/default.sh
@@ -125,7 +125,7 @@ if [ -z "$TMUX_POWERLINE_RIGHT_STATUS_SEGMENTS" ]; then
 		"pwd 89 211"
 		#"macos_notification_count 29 255"
 		#"mailcount 9 255"
-		"now_playing 234 37"
+		# "now_playing 234 37"
 		#"cpu 240 136"
 		"load 237 167"
 		#"tmux_mem_cpu_load 234 136"